requirements.txt
pydantic
typing
python-dotenv
pydantic-settings
grpcio-tools
grpcio